Updates tests and support for python 2/3
This is a big diff, in part because tests were quite flaky. A lot of
work went into shoring those up to ensure confidence that nothing was
broken in this update.

This introduces a new dependency of `requests` to make dealing with http
a bit easier between 2/3 (issue was posting gzip content). If dropping
Python 2 in a future release, the we can probably go back to just using
the standard http client since the use case is quite simple.